+++ styleguide.config.js
@@ -1,11 +1,4 @@
 const path = require('path');
-const craWebpack = require('react-scripts/config/webpack.config'); // eslint-disable-line
-
-const getWebpack = () => {
-  const config = craWebpack('development');
-  config.resolve.alias['state-template'] = path.join(path.join(__dirname, 'src/lib/'));
-  return config;
-};
 
 module.exports = {
   skipComponentsWithoutExample: true,
@@ -18,7 +11,6 @@ module.exports = {
   styleguideComponents: {
     Wrapper: path.join(__dirname, 'src/styleguidist/Wrapper'),
   },
-  webpackConfig: getWebpack(),
   getComponentPathLine(componentPath) {
     const name = path.basename(componentPath, '.js');
     return `import { ${name} } from 'state-template';`;